pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
mysql-connector = "^2.2.9"
XlsxWriter = "^3.0.8"
bs4 = "^0.0.1"
requests = "^2.28.2"
collections2 = "^0.3.0"
DateTime = "^5.0"
mysql-connector-python = "^8.0.32"
google-search = "^1.1.1"
google = "^3.0.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"